+++ backend/schemas/book_schema.py
@@ -6,7 +6,7 @@ class BookSchema(Schema):
     title = fields.Str(required=True, validate=validate.Length(min=1, error="Title cannot be empty"))
     author = fields.Str(required=True, validate=validate.Length(min=1, error="Author cannot be empty"))
     cover = fields.Url(allow_none=True)
-    rating = fields.Int(validate=validate.Range(min=0, max=5, error="Rating must be between 0 and 5"))
+    rating = fields.Float(validate=validate.Range(min=0, max=5, error="Rating must be between 0 and 5"))
     pages = fields.Int(allow_none=True)
     genre = fields.Str(allow_none=True)
     status = fields.Str(validate=validate.OneOf(["want-to-read", "reading", "completed"], error="Status must be one of: want-to-read, reading, completed"))
